% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima.R
\name{arima.aic.bic}
\alias{arima.aic.bic}
\title{获得备选模型的aic值}
\usage{
arima.aic.bic(model.list, seasonal = TRUE)
}
\arguments{
\item{model.list}{T: list<Arima> 备选模型列表}

\item{seasonal}{T: logical 是否提取季节性（P, D, Q）}
}
\value{
T: data.frame[col.names <- c("model", "aic", "bic", "p", "d", "q")]
}
\description{
获得备选模型的aic值
}
