% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima.R
\name{arima.arg}
\alias{arima.arg}
\title{生成 ARIMA 模型的参数列表}
\usage{
arima.arg(order, seasonal = NULL, period = 12)
}
\arguments{
\item{order}{T: vector<int> ARIMA 模型的阶数参数，包括 p, d, q 三个参数。}

\item{seasonal}{T: vector<int> 季节性参数，包括 P, D, Q 三个参数。}

\item{period}{T: int 季节性周期，即数据的周期性间隔，默认为 12（即每年 12 个观测值）。}
}
\value{
T: list 返回 ARIMA 模型的参数列表，包括 `order` 和 `seasonal` 两个部分。
}
\description{
该函数用于生成 ARIMA 模型的参数列表，包括 `order` 和 `seasonal` 两个部分。
`order` 部分包含 ARIMA 模型的阶数参数，`seasonal` 部分包含季节性参数。
}
\examples{
# 生成 ARIMA 模型的参数列表
arima.arg(c(1, 1, 1), c(1, 1, 1), period = 12)
}
