% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima.R
\name{arima.coef.se.list.builder}
\alias{arima.coef.se.list.builder}
\title{通过取得的arima.aic.bic表的结果，生成表中提到的全部模型的系数查看列表}
\usage{
arima.coef.se.list.builder(model.name.vector)
}
\arguments{
\item{model.name.vector}{T: vector<char> 备选模型的名称列表}
}
\value{
T: list 返回包含所有模型系数查看列表的列表
}
\description{
该函数用于通过取得的arima.aic.bic表的结果，生成表中提到的全部模型的系数查看列表
}
\examples{
# 生成模型系数查看列表
coef.se.list <- arima.coef.se.list.builder(model.name.vector)
}
