% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima.R
\name{arima.coef.se}
\alias{arima.coef.se}
\title{提取ARIMA模型系数和标准误}
\usage{
arima.coef.se(model)
}
\arguments{
\item{model}{T: arima 模型对象}
}
\value{
T: data.frame 返回包含系数及其标准误的表格
}
\description{
提取ARIMA模型系数和标准误
}
\details{
**报错**

`Warning message: In sqrt(diag(x$var.coef)) : NaNs produced` -> 您选择的p,d,q,P,D,Q不合理
}
\examples{
tidy_model <- arima.coef.se(model.list$fit.2)
}
