% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima.R
\name{arima.convert.to.model.list}
\alias{arima.convert.to.model.list}
\title{将 ARIMA 模型参数列表转换为模型对象列表}
\usage{
arima.convert.to.model.list(arg.list, data)
}
\arguments{
\item{arg.list}{T: list 包含 ARIMA 模型参数的列表。}

\item{data}{T: ts 时间序列数据。}
}
\value{
T: list 返回 ARIMA 模型对象列表。
}
\description{
该函数用于将 ARIMA 模型的参数列表转换为 ARIMA 模型对象列表。适用于季节性生成
}
\examples{
# 生成 ARIMA 模型参数列表
arg.list <- list(
  arima.arg(c(1, 1, 1), c(1, 1, 1), period = 12),
  arima.arg(c(1, 1, 1), c(1, 1, 2), period = 12)
)
}
