% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima.R
\name{arima.forecast.data.frame}
\alias{arima.forecast.data.frame}
\title{获得拟合与预测序列}
\usage{
arima.forecast.data.frame(train, check, fit, forecast, h = 12, level = 95)
}
\arguments{
\item{train}{T: data.frame[required.col.names <-  c(date: Date)] 传入的训练数据集}

\item{check}{T: data.frame[required.col.names <-  c(date: Date)] 传入的验证数据集}

\item{fit}{T: Arima 传入的模型}

\item{forecast}{T: charactor 指定要被预测的列名}

\item{h}{T: num 向后预测的个数}

\item{level}{T: num 置信区间\%}
}
\value{
T: data.frame[col.names <- c("date", "actual", "forecast", "lo", "hi")]
}
\description{
获得拟合与预测序列
}
