% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima.R
\name{arima.noise.test}
\alias{arima.noise.test}
\title{TODO: 执行 ARIMA 模型的白噪声检验}
\usage{
arima.noise.test(observed)
}
\arguments{
\item{observed}{T: numeric 向量，观察值序列}
}
\value{
T: list 返回白噪声检验的结果，包括统计量和p值
}
\description{
对给定的残差序列进行白噪声检验，检查残差是否为白噪声。
}
\examples{
noise_test <- arima.noise.test(model.list$fit.2$residuals)
}
