% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima.R
\name{arima.plot.acf.pacf}
\alias{arima.plot.acf.pacf}
\title{绘制 ACF 和 PACF 图}
\usage{
arima.plot.acf.pacf(data, observe)
}
\arguments{
\item{data}{T: data.frame 包含时间序列数据的数据框}

\item{observe}{T: char 需要绘制 ACF 和 PACF 的列名（必须是数值型列）}
}
\value{
ggplot 对象，包含 ACF 和 PACF 图
}
\description{
本函数用于绘制指定数据列的自相关函数（ACF）和偏自相关函数（PACF）图，
并使用 `patchwork` 进行排版。
}
\examples{
# 示例数据
file.time <- data.frame(
  date = seq(as.Date("2020-01-01"), by = "month", length.out = 36),
  incidence.rate = rnorm(36, mean = 50, sd = 10)
)

# 绘制 ACF 和 PACF 图
p <- arima.plot.acf.pacf(file.time, "incidence.rate")
}
