% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima.R
\name{arima.plot.forecast.comparison}
\alias{arima.plot.forecast.comparison}
\title{绘制实际数据与拟合&预测数据的对比折线图}
\usage{
arima.plot.forecast.comparison(fitted.data.frame, title)
}
\arguments{
\item{fitted.data.frame}{T: data.frame[required.col.names <-  c("date", "incidence.rate", "lo", "hi", )]}

\item{title}{T: charactor 图像的标题}
}
\value{
T: list<Any> 实际数据与拟合&预测数据的对比折线图元数据列表
}
\description{
绘制实际数据与拟合&预测数据的对比折线图
}
