% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima.R
\name{arima.prediction.performance}
\alias{arima.prediction.performance}
\title{获得ARIMA模型预测评价指标}
\usage{
arima.prediction.performance(fit, check, train, forecast)
}
\arguments{
\item{fit}{T: arima 模型对象}

\item{check}{T: data.frame，验证集数据，包含实际数据和预测数据}

\item{train}{T: data.frame，训练集数据，包含实际数据和预测数据}

\item{forecast}{T: data.frame，包含拟合的预测结果}
}
\value{
T: list<matrix> 函数将返回包括 ME, RMSE, MAE, MPE, MAPE, MASE, ACF1 等误差度量的矩阵列表
}
\description{
包括 ME, RMSE, MAE, MPE, MAPE, MASE, ACF1 等误差度量用于评价ARIMA模型的性能。
}
\examples{
error_measures <- arima.prediction.performance(model.list$fit.2, check_data, train_data, forecast_data)
}
