% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima.R
\name{arima.split.sample}
\alias{arima.split.sample}
\title{拆分样本}
\usage{
arima.split.sample(data, cut.point = NULL, percentage = c(80, 20))
}
\arguments{
\item{data}{T: data.frame[required.col.names <- c(date: Date)] 被拆分的数据集，date列必须是Date类型}

\item{cut.point}{T: Date char 分割点，必须是Date格式的字符串}

\item{percentage}{T: vector<num, num> 分割比例，代表训练集和验证集的百分点}
}
\description{
拆分样本
}
