% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima.R
\name{arima_fitted_with_actual}
\alias{arima_fitted_with_actual}
\title{获得拟合序列}
\usage{
arima_fitted_with_actual(fit, train, forecast)
}
\arguments{
\item{fit}{T: Arima 传入的模型}

\item{train}{train T: data.frame[required.col.names <-  c(date: Date)] 传入的训练数据集}

\item{forecast}{T: charactor 指定要被预测的列名}
}
\value{
T: data.frame[col.names <- c("date", "actual", "forecast", "lo", "hi")]
}
\description{
获得拟合序列
}
