% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima.R
\name{arima_forecast_with_ci}
\alias{arima_forecast_with_ci}
\title{获得模型预测序列}
\usage{
arima_forecast_with_ci(check, fit, forecast, level = 95)
}
\arguments{
\item{check}{T: data.frame[required.col.names <-  c(date: Date)] 传入的验证数据集}

\item{fit}{T: Arima 传入的模型}

\item{forecast}{T: charactor 指定要被预测的列名}

\item{level}{T: num 置信区间\%}

\item{h}{T: num 向后预测的个数}
}
\value{
T: data.frame[col.names <- c("date", "actual", "forecast", "lo", "hi")]
}
\description{
获得模型预测序列
}
