% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lstm.R
\name{lstm.convert.scaled.data}
\alias{lstm.convert.scaled.data}
\title{生成比例数据}
\usage{
lstm.convert.scaled.data(data)
}
\arguments{
\item{data}{T: vector<double> ~> 输入的数据}
}
\value{
T: vector<double> ~> 比例类型的向量，值范围`[0,1]`
}
\description{
生成的比例数据是原数据在传入序列范围内的百分比位置，越靠近低点值越小。
}
