% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lstm.R
\name{lstm.create.dataset}
\alias{lstm.create.dataset}
\title{数据拆分为X和Y}
\usage{
lstm.create.dataset(set, look_back)
}
\arguments{
\item{set}{T: vector<double> ~> 输入的数据}

\item{look_back}{T: num ~> 回溯的数据步长：如果在时间序列中，当前值的预测值与前n个数有关，那么回溯的数据步长为n}
}
\value{
T: list<array> ~>返回一个list，其中包含训练窗口数组（windows）和验证数据数组（confirms）
}
\description{
通过一个向量生成一个训练窗口集合以及一个验证集
}
