% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima.R
\name{my.checkresiduals}
\alias{my.checkresiduals}
\title{TODO:检查残差}
\usage{
my.checkresiduals(model.list)
}
\arguments{
\item{model.list}{T: list<Arima> 备选模型列表}
}
\value{
T: list<Any> 有关于图像原数据的列表
}
\description{
TODO:检查残差
}
