% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile.R
\name{quantile_by}
\alias{quantile_by}
\title{批量读取csv文件方法}
\usage{
quantile_by(colname, data, probs)
}
\arguments{
\item{data}{T: data.frame 数据框}

\item{probs}{T: vector<double> 分位数向量}

\item{col.names}{T: vector<charactor> 列名}
}
\value{
T: factor 返回一个因子，表示每个数据点所属的分位数区间
}
\description{
批量读取csv文件方法
}
