% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_file.R
\name{read.patch.csv}
\alias{read.patch.csv}
\title{批量读取csv文件方法}
\usage{
read.patch.csv(file.names)
}
\arguments{
\item{file.names}{T: vector<charactor> 文件名}
}
\value{
T: data.frame[col.names <- c(...)] 合并文件的数据框
}
\description{
批量读取csv文件方法
}
