% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_file.R
\name{read.patch.xport}
\alias{read.patch.xport}
\title{批量读取xport文件方法}
\usage{
read.patch.xport(file.names)
}
\arguments{
\item{file.names}{T: vector<charactor> 文件名，要求传入的文件拥有相同数量的变量和相同的变量名称}
}
\value{
T: data.frame[col.names <- c(...)] 合并文件的数据框
}
\description{
批量读取xport文件方法
}
